requirements.txt
tensorflow>=2.12.0
flask>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
